+++ trunk/cpsycle/audio/src/duplicator.c
@@ -108,10 +108,10 @@ void dispose(psy_audio_Duplicator* self)
 
 void initparameters(psy_audio_Duplicator* self)
 {
-	uintptr_t gbp;
+	int gbp;
 
 	psy_table_init(&self->parameters);
-	for (gbp = 0; gbp < psy_audio_duplicatormap_numoutputs(&self->map); ++gbp) {
+	for (gbp = 0; gbp < (int)psy_audio_duplicatormap_numoutputs(&self->map); ++gbp) {
 		char name[128];
 		psy_audio_IntMachineParam* param;
 		psy_audio_DuplicatorOutput* output;
@@ -168,7 +168,7 @@ psy_List* sequencerinsert(psy_audio_Duplicator* self, psy_audio_PatternNode* eve
 					psy_audio_PatternEntry* patternentry;					
 					psy_audio_PatternEntry* newentry;
 					int note;
-					int mapchannel;
+					uintptr_t mapchannel;
 					patternentry = psy_audio_patternnode_entry(p);
 					newentry = psy_audio_patternentry_clone(patternentry);
 					note = psy_audio_patternentry_front(patternentry)->note;
