+++ trunk/cpsycle/audio/src/machineinfo.h
@@ -31,7 +31,7 @@ typedef struct {
 	/// module path
 	char* modulepath;
 	/// shellidx
-	int shellidx;
+	uintptr_t shellidx;
 	char* helptext;
 } psy_audio_MachineInfo;
 
@@ -52,7 +52,7 @@ void machineinfo_set(psy_audio_MachineInfo*,
 		int16_t plugversion,
 		int type,		
 		const char* modulepath,
-		int shellidx,
+		uintptr_t shellidx,
 		const char* helptext);
 struct CMachineInfo;		
 void machineinfo_setnativeinfo(psy_audio_MachineInfo*,
